Separate URLType into its own file
