+++ fastapi_hypermodel/url_for.py
@@ -20,11 +20,10 @@ from pydantic_core import CoreSchema
 from typing_extensions import Self
 
 from fastapi_hypermodel.hypermodel import (
-    URL_TYPE_SCHEMA,
     AbstractHyperField,
     HasName,
-    UrlType,
 )
+from fastapi_hypermodel.url_type import URL_TYPE_SCHEMA, UrlType
 from fastapi_hypermodel.utils import get_route_from_app, resolve_param_values
